+++ crates/model/src/events/order/modify_rejected.rs
@@ -222,7 +222,7 @@ impl OrderEvent for OrderModifyRejected {
     }
 
     fn reconciliation(&self) -> bool {
-        false
+        self.reconciliation
     }
 
     fn price(&self) -> Option<Price> {
@@ -323,6 +323,9 @@ impl OrderEvent for OrderModifyRejected {
 
     fn ts_init(&self) -> UnixNanos {
         self.ts_init
+    }
+    fn causation_id(&self) -> Option<UUID4> {
+        self.causation_id
     }
 }