+++ crates/model/src/events/order/triggered.rs
@@ -217,7 +217,7 @@ impl OrderEvent for OrderTriggered {
     }
 
     fn reconciliation(&self) -> bool {
-        false
+        self.reconciliation
     }
 
     fn price(&self) -> Option<Price> {
@@ -318,6 +318,9 @@ impl OrderEvent for OrderTriggered {
 
     fn ts_init(&self) -> UnixNanos {
         self.ts_init
+    }
+    fn causation_id(&self) -> Option<UUID4> {
+        self.causation_id
     }
 }